backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
sqlmodel==0.0.22
asyncpg==0.30.0

# Data processing
pandas==2.2.3
openpyxl==3.1.5

# Environment variables
python-dotenv==1.0.1

# CORS
python-multipart==0.0.18

# AI (Phase 5)
anthropic==0.42.0
